fotoUpload works
